Make the application production runnable
