+++ app/jobs/moh_report_job.rb
@@ -6,7 +6,7 @@ class MohReportJob
 
   def perform
     report_years = Reports::Moh::ReportUtils.report_years.reverse.first(2)
-    report_years[].each do |year|
+    report_years.each do |year|
       Report.find_or_create_by(name: 'moh_haematology', year:).update(
         data: Reports::MohService.generate_haematology_report(year)
       )
