+++ config/environments/production.rb
@@ -52,8 +52,8 @@ Rails.application.configure do
   # config.cache_store = :mem_cache_store
 
   # Use a real queuing backend for Active Job (and separate queues per environment).
-  # config.active_job.queue_adapter     = :resque
-  # config.active_job.queue_name_prefix = "mlab_api_production"
+  config.active_job.queue_adapter = :sidekiq
+  config.active_job.queue_name_prefix = "mlab_api_prod"
 
   config.action_mailer.perform_caching = false
