+++ iblis_migration/iblis/culture_observations.rb
@@ -1,52 +1,53 @@
 # frozen_string_literal: true
-
+require_relative '../iblis/iblis_service/drug_organism_service.rb'
+require_relative '../iblis/iblis_service/load_client_service.rb'
+require_relative '../iblis/iblis_service/measure_service.rb'
+require_relative '../iblis/iblis_service/status_service.rb'
 Rails.logger = Logger.new($stdout)
 
-def iblis_drug_susceptibility(offset, limit, creator)
+def iblis_cs_observations(offset, limit, creator)
   Iblis.find_by_sql("
     SELECT
       id,
       user_id AS creator,
       test_id,
-      organism_id,
-      drug_id,
-      zone,
-      interpretation,
+      observation AS description,
       created_at AS created_date,
       updated_at AS updated_date,
+      created_at AS observation_datetime,
       user_id AS updated_by,
-      CASE WHEN deleted_at IS NOT NULL THEN 1 ELSE 0 END AS voided,
-      CASE WHEN deleted_at IS NOT NULL THEN user_id ELSE NULL END AS voided_by,
+      0 AS voided,
+      NULL AS voided_by,
       NULL AS voided_reason,
-      CASE WHEN deleted_at IS NOT NULL THEN deleted_at ELSE NULL END AS voided_date
+      NULL AS voided_date
       FROM
-      drug_susceptibility
+      culture_worksheet
     LIMIT #{limit} OFFSET #{offset}
   ")
 end
 
-def iblis_drug_susceptibility_count
+def iblis_cs_observations_count
   Iblis.find_by_sql("
     SELECT
       count(*) AS count
     FROM
-    drug_susceptibility
+      culture_worksheet
   ")[0]
 end
 ActiveRecord::Base.connection.execute("SET sql_mode='NO_ZERO_DATE'")
 creator = User.first.id
 Rails.logger.info('Starting to process....')
-total_records = iblis_drug_susceptibility_count.count
+total_records = iblis_cs_observations_count.count
 batch_size = 50_000
 offset = 0
 count = total_records
 loop do
   ActiveRecord::Base.connection.execute("SET FOREIGN_KEY_CHECKS=0")
-  records = iblis_drug_susceptibility(offset, batch_size, creator)
+  records = iblis_cs_observations(offset, batch_size, creator)
   break if records.empty?
 
-  Rails.logger.info("Processing batch #{offset} of #{total_records}: Remaining - #{count} --Drug Susceptiblity-- step(9 of 10)")
-  DrugSusceptibility.upsert_all(records.map(&:attributes), returning: false) unless records.empty?
+  Rails.logger.info("Processing batch #{offset} of #{total_records}: Remaining - #{count} --Culture Observations-- step(9 of 10)")
+  CultureObservation.upsert_all(records.map(&:attributes), returning: false) unless records.empty?
   offset += batch_size
   count -= batch_size
   ActiveRecord::Base.connection.execute("SET FOREIGN_KEY_CHECKS=1")
