+++ load_iblis_data.sh
@@ -21,38 +21,38 @@ read steps
 echo "==== starting migrations ===="
 if [ -z "$steps" ]; then
     # Run all steps if no input is provided
-    rails r bin/iblis/init_migration.rb &&
-    rails r bin/iblis/load_users_data.rb &&
-    rails r bin/iblis/load_clients.rb &&
-    rails r bin/iblis/meta_data.rb &&
-    rails r bin/iblis/migrate_intruments.rb &&
-    rails r bin/iblis/load_encounter_types.rb &&
-    rails r bin/iblis/load_facility_data.rb &&
-    rails r bin/iblis/load_encounter.rb &&
-    rails r bin/iblis/load_orders.rb &&
-    rails r bin/iblis/load_test.rb &&
-    rails r bin/iblis/load_test_results.rb &&
-    rails r bin/iblis/culture_observations.rb &&
-    rails r bin/iblis/drug_susceptibilities.rb &&
-    rails r bin/iblis/load_order_print_trail.rb
+    RAILS_ENV=production rails r iblis_migration/iblis/init_migration.rb &&
+    RAILS_ENV=production rails r iblis_migration/iblis/load_users_data.rb &&
+    RAILS_ENV=production rails r iblis_migration/iblis/load_clients.rb &&
+    RAILS_ENV=production rails r iblis_migration/iblis/meta_data.rb &&
+    RAILS_ENV=production rails r iblis_migration/iblis/migrate_intruments.rb &&
+    RAILS_ENV=production rails r iblis_migration/iblis/load_encounter_types.rb &&
+    RAILS_ENV=production rails r iblis_migration/iblis/load_facility_data.rb &&
+    RAILS_ENV=production rails r iblis_migration/iblis/load_encounter.rb &&
+    RAILS_ENV=production rails r iblis_migration/iblis/load_orders.rb &&
+    RAILS_ENV=production rails r iblis_migration/iblis/load_test.rb &&
+    RAILS_ENV=production rails r iblis_migration/iblis/load_test_results.rb &&
+    RAILS_ENV=production rails r iblis_migration/iblis/culture_observations.rb &&
+    RAILS_ENV=production rails r iblis_migration/iblis/drug_susceptibilities.rb &&
+    RAILS_ENV=production rails r iblis_migration/iblis/load_order_print_trail.rb
 else
     # Run selected steps based on user input
     for step in $steps; do
         case $step in
-            1) rails r bin/iblis/init_migration.rb ;;
-            2) rails r bin/iblis/load_users_data.rb ;;
-            3) rails r bin/iblis/load_clients.rb ;;
-            4) rails r bin/iblis/meta_data.rb ;;
-            5) rails r bin/iblis/migrate_intruments.rb ;;
-            6) rails r bin/iblis/load_encounter_types.rb ;;
-            7) rails r bin/iblis/load_facility_data.rb ;;
-            8) rails r bin/iblis/load_encounter.rb ;;
-            9) rails r bin/iblis/load_orders.rb ;;
-            10) rails r bin/iblis/load_test.rb ;;
-            11) rails r bin/iblis/load_test_results.rb ;;
-            12) rails r bin/iblis/culture_observations.rb ;;
-            13) rails r bin/iblis/drug_susceptibilities.rb;;
-            14) rails r bin/iblis/load_order_print_trail.rb ;;
+            1) RAILS_ENV=production rails r iblis_migration/iblis/init_migration.rb ;;
+            2) RAILS_ENV=production rails r iblis_migration/iblis/load_users_data.rb ;;
+            3) RAILS_ENV=production rails r iblis_migration/iblis/load_clients.rb ;;
+            4) RAILS_ENV=production rails r iblis_migration/iblis/meta_data.rb ;;
+            5) RAILS_ENV=production rails r iblis_migration/iblis/migrate_intruments.rb ;;
+            6) RAILS_ENV=production rails r iblis_migration/iblis/load_encounter_types.rb ;;
+            7) RAILS_ENV=production rails r iblis_migration/iblis/load_facility_data.rb ;;
+            8) RAILS_ENV=production rails r iblis_migration/iblis/load_encounter.rb ;;
+            9) RAILS_ENV=production rails r iblis_migration/iblis/load_orders.rb ;;
+            10) RAILS_ENV=production rails r iblis_migration/iblis/load_test.rb ;;
+            11) RAILS_ENV=production rails r iblis_migration/iblis/load_test_results.rb ;;
+            12) RAILS_ENV=production rails r iblis_migration/iblis/culture_observations.rb ;;
+            13) RAILS_ENV=production rails r iblis_migration/iblis/drug_susceptibilities.rb;;
+            14) RAILS_ENV=production rails r iblis_migration/iblis/load_order_print_trail.rb ;;
             *) echo "Invalid step number: $step" ;;
         esac
     done
